requirements.txt
paddlenlp==2.3.1
paddlepaddle==2.3.0
absl-py
h5py